requirements.txt
affine==2.2.0
albumentations==0.4.5
attrs==17.4.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
click-plugins==1.0.3
cligj==0.4.0
cycler==0.10.0
decorator==4.2.1
idna==2.6
kiwisolver==1.0.1
matplotlib==2.2.2
networkx==2.1
numpy==1.14.5
opencv-python==3.4.0.12
Pillow==5.0.0
pyparsing==2.2.0
pyproj==2.4.2
python-dateutil==2.7.2
pytz==2018.3
PyWavelets==0.5.2
rasterio==1.0.24
requests==2.18.4
scikit-image==0.13.1
scikit-learn==0.19.1
seaborn==0.10.0
Shapely==1.6.4.post1
six==1.11.0
scipy==1.1.0
snuggs==1.4.1
tqdm==4.20.0
urllib3==1.22
nudged==0.3.1
opencv-contrib-python==3.4.1.15
firebase-admin==2.10.0
tensorflow-addons==0.6.0
tensorflow-gpu==2.0.0
Keras==2.2.4
Keras-Applications==1.0.8
keras-resnet==0.1.0
pandas==0.23.4
h5py==2.8.0
geopandas==0.4.0
sknw==0.13
segmentation-models==1.0.1
